third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-in-http-header-control-expected.txt

CONSOLE WARNING: An iframe which has both allow-scripts and allow-same-origin for its sandbox attribute can escape its sandboxing.
CONSOLE MESSAGE: Script executed in iframe.
CONSOLE MESSAGE: PASS: Iframe was not in a unique origin
